피드백3. CommonUtils 다른 성격의 메소드 분리 피드백4. AuthServiceImpl.login() 1) HttpServletRequest 서비스단까지 가져오지말기 2) Null 체크보단 Optional 사용 3) Fail-Fast 피드백5. AuthServiceImpl.getAuthenticationBean 1) String.valudOf 불필요 2) 토큰 만료데이터 처리 방법 3) Setter 처리
