+++ src/main/java/com/book/account/auth/service/AuthService.java
@@ -1,7 +1,5 @@
 package com.book.account.auth.service;
 
-import javax.servlet.http.HttpServletRequest;
-
 import com.book.account.auth.model.dto.AuthenticationBean;
 import com.book.account.auth.model.dto.LoginDto;
 import com.book.account.auth.model.dto.RegisterDto;
@@ -11,7 +9,7 @@ public interface AuthService {
     /**
      * 로그인
      */
-    AuthenticationBean login(LoginDto loginDto, HttpServletRequest request);
+    AuthenticationBean login(LoginDto loginDto, String ip);
 
     /**
      * 사용자 등록
