+++ src/main/java/com/book/account/module/model/Module.java
@@ -7,7 +7,6 @@ import javax.persistence.Table;
 import com.book.account.common.model.BaseEntity;
 import com.book.account.module.model.dto.ModuleUpdateDto;
 
-import org.springframework.util.NumberUtils;
 import org.springframework.util.StringUtils;
 
 import lombok.AllArgsConstructor;
