+++ src/main/java/com/book/account/util/CommonUtils.java
@@ -1,41 +1,9 @@
 package com.book.account.util;
 
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-
 import javax.servlet.http.HttpServletRequest;
 
 public class CommonUtils {
     
-    public static String encBySha256(String data) {
-        String returnVal = "";
-        try {
-            MessageDigest md = MessageDigest.getInstance("SHA-256");
-            md.update(data.getBytes());
-
-            byte byteData[] = md.digest();
-            StringBuilder sb = new StringBuilder();
-
-            for (int i = 0; i < byteData.length; i++) {
-                sb.append(Integer.toString((byteData[i] & 0xff) + 0x100, 16).substring(1));
-            }
-
-            StringBuilder hexString = new StringBuilder();
-            for (int i = 0; i < byteData.length; i++) {
-                String hex = Integer.toHexString(0xff & byteData[i]);
-                if (hex.length() == 1) {
-                    hexString.append('0');
-                }
-                hexString.append(hex);
-            }
-
-            returnVal = hexString.toString();
-        } catch (NoSuchAlgorithmException e) {
-
-        }
-        return returnVal;
-    }
-
     /**
      * CLIKENT IP 주소
      */
