+++ src/main/java/com/book/account/util/EncUtils.java
@@ -1,0 +1,35 @@
+package com.book.account.util;
+
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+
+public class EncUtils {
+    public static String encBySha256(String data) {
+        String returnVal = "";
+        try {
+            MessageDigest md = MessageDigest.getInstance("SHA-256");
+            md.update(data.getBytes());
+
+            byte byteData[] = md.digest();
+            StringBuilder sb = new StringBuilder();
+
+            for (int i = 0; i < byteData.length; i++) {
+                sb.append(Integer.toString((byteData[i] & 0xff) + 0x100, 16).substring(1));
+            }
+
+            StringBuilder hexString = new StringBuilder();
+            for (int i = 0; i < byteData.length; i++) {
+                String hex = Integer.toHexString(0xff & byteData[i]);
+                if (hex.length() == 1) {
+                    hexString.append('0');
+                }
+                hexString.append(hex);
+            }
+
+            returnVal = hexString.toString();
+        } catch (NoSuchAlgorithmException e) {
+
+        }
+        return returnVal;
+    }
+}
